% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{check_simulation_requirement}
\alias{check_simulation_requirement}
\title{Check simulation requirements}
\usage{
check_simulation_requirement()
}
\value{
A logical value indicating whether all the required packages are loaded.
}
\description{
This function checks if the required packages for running simulations are loaded.
It checks for the presence of the "doSNOW", "greybox", "mvtnorm", and "extraDistr" packages.
}
