% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_check.R
\name{convergence_check}
\alias{convergence_check}
\title{Check convergence of a multiplex network model}
\usage{
convergence_check(
  multip2fit,
  params,
  custom_params = "",
  plot = c("traceplot", "running", "geweke", "density"),
  plot_all = FALSE,
  file = "ggmcmc-output.pdf"
)
}
\arguments{
\item{multip2fit}{the fitted MultiP2Fit object.}

\item{params}{A character string indicating which parameters to plot. Must be one of "fixed" (all the fixed parameters), "random"(actor random effects), or "custom" (use the custom_params argument)}

\item{custom_params}{A character string indicating a custom regular expression to match parameter names to plot, as given by a character vector or a regular expression, i.e., the "family" parameter in ggmcmc::ggs.}

\item{plot}{A character vector containing the names of the desired plots.m Default is c("traceplot", "running", or "geweke")}

\item{plot_all}{A logical indicating whether to plot all types of plots. If TRUE, the plot argument is ignored, instead all plots are generated.}

\item{file}{A character string indicating the output file name for the plots, default = "ggmcmc-output.pdf"}
}
\description{
This function checks the convergence of a fitted Stan model for a multiplex network using the ggmcmc package.
}
