% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_summary}
\alias{create_summary}
\title{Create summary for a model object}
\usage{
create_summary(model_obj)
}
\arguments{
\item{model_obj}{The model object for which the summary is to be created.}
}
\value{
A list containing the fixed and random effects summaries, along with parameter labels.
}
\description{
This function creates a summary for a given model object. It extracts fixed and random effects summaries
and returns them along with parameter labels.
}
