% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Calculate descriptive statistics for a network}
\usage{
descriptive_stats(networks)
}
\arguments{
\item{network}{An igraph object representing the network.}
}
\value{
A numeric vector of length 3. The first element is the edge density, the second element is the reciprocity, and the third element is the transitivity.
}
\description{
This function calculates the edge density, reciprocity, and transitivity of a network.
}
\examples{
# TODO: Add examples of usage
}
