% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{descriptive_stats_prior}
\alias{descriptive_stats_prior}
\title{Calculate network descriptive statistics for a given adjacency matrix}
\usage{
descriptive_stats_prior(X)
}
\arguments{
\item{X}{An adjacency matrix representing a network}
}
\value{
A numeric vector containing the calculated descriptive statistics
}
\description{
This function takes an adjacency matrix as input and calculates three descriptive statistics:
\itemize{
\item Edge density: the proportion of possible edges that are present in the network
\item Reciprocity: the proportion of edges that are reciprocated (i.e., both nodes are connected to each other)
\item Transitivity: the proportion of triangles in the network (i.e., the tendency for nodes to form clusters)
}
}
