% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dyads_to_matrix_1d}
\alias{dyads_to_matrix_1d}
\title{Reconstruct the Adjacency Matrix from Dyad-wise Results}
\usage{
dyads_to_matrix_1d(y, n)
}
\arguments{
\item{y}{A vector of dyad-wise results. Each element should be in {1,2,3,4}.}

\item{n}{The number of actors.}
}
\value{
An adjacency matrix reconstructed from the dyad-wise results.
}
\description{
Given a vector of dyad-wise results \code{y} in {1,2,3,4}, and the number of actors \code{n},
this function reconstructs the adjacency matrix.
}
\examples{
# replace with a valid example

}
