% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dyads_to_matrix_list}
\alias{dyads_to_matrix_list}
\title{Convert Dyadic Data to List of Adjacency Matrices}
\usage{
dyads_to_matrix_list(dyad_df, n, t, dep_lab, network_type)
}
\arguments{
\item{dyad_df}{A B x N dataframe where each row is a network in dyad form.}

\item{n}{Number of actors per network.}

\item{t}{Number of layers in the networks.}

\item{network_type}{Type of network to return. Options are "adj" for adjacency matrix (default), "igraph" for igraph object, or "network" for network object.}
}
\value{
A list of B lists of t adjacency matrices.
}
\description{
This function takes a dataframe of dyadic data and converts it to a list of adjacency matrices.
Each row of the dataframe is a network in dyad form (1 x N) with B total such networks. N is the number of dyads in the network.
The function returns a list of B lists of t adjacency matrices representing the B multiplex networks.
}
