% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dyads_to_matrix_nd}
\alias{dyads_to_matrix_nd}
\title{Convert Dyadic Data to List of Adjacency Matrices for Multiple Networks}
\usage{
dyads_to_matrix_nd(y_nd, t, n)
}
\arguments{
\item{y_nd}{A vector of dyad-wise results for multiple networks. Each element should be in {1,...,2^(2*t)}.}

\item{t}{The total number of networks.}

\item{n}{The number of actors.}
}
\value{
A list of \code{t} adjacency matrices.
}
\description{
Given a vector \code{y_nd} in {1,...,2^(2*t)}^N, total number of networks \code{t}, and number of actors \code{n},
this function returns a list of \code{t} networks in adjacency matrix form.
}
\examples{
# replace with a valid example

}
