% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{extract_draws}
\alias{extract_draws}
\title{Extracts draws from a Stan model fit}
\usage{
extract_draws(fit, parameter)
}
\arguments{
\item{fit}{A Stan model fit object.}

\item{parameter}{The name of the parameter to extract draws for, options are "mu", "rho", "cross_mu", "cross_rho", "alpha", "beta", "Sigma".}
}
\value{
A vector of draws for the specified parameter.
}
\description{
This function takes a Stan model fit object and a parameter name, and returns
the draws for that parameter from the fit.
}
\examples{
#draws <- extract_draws(stan_fit, "mu")

}
