% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{extract_network_draws}
\alias{extract_network_draws}
\title{Extract simulated network outcome from the (prior) posterior of a fitted stan object}
\usage{
extract_network_draws(fit, sim_num, as_adjacency = TRUE)
}
\arguments{
\item{fit}{rstan fit object: fitted stan object}

\item{sim_num}{integer: number of simulations to extract, counting from the tail of the posterior draws}

\item{as_adjacency}{logical: whether to return the simulated network outcome as an adjacency matrix, default is TRUE}
}
\value{
the (prior) posterior draws of the simulated network outcome
}
\description{
Extract simulated network outcome from the (prior) posterior of a fitted stan object
}
