% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fit.MultiP2Fit}
\alias{fit.MultiP2Fit}
\title{fit function for "MultiP2Fit" class}
\usage{
\method{fit}{MultiP2Fit}(
  model_obj,
  chains = 4,
  iter = 200,
  warmup = floor(iter/2),
  thin = 1,
  seed = sample.int(.Machine$integer.max, 1),
  stan_file = "multiplex_p2_low_mem.stan",
  prior_sim = FALSE,
  ...
)
}
\arguments{
\item{model_obj}{An object of class "MultiP2Fit"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
fit function for "MultiP2Fit" class
}
