% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{fit_stan}
\alias{fit_stan}
\title{Format covariates data for stan estimation}
\usage{
fit_stan(stan_data, chains = 4, warmup = 4000, thin = 1, parallel = TRUE)
}
\arguments{
\item{stan_data}{list: network and covariates data for the stan program}

\item{chains}{numeric: number of MCMC chains, default = 4}

\item{warmup}{numeric: number of warmup MCMC steps, iter = 2 * warmup, default = 4000}

\item{thin}{numeric: TODO}

\item{parallel}{boolean: whether or not to parallelize the chains}
}
\value{
stan_fit object: results of the fit
}
\description{
Format covariates data for stan estimation
}
