% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdata_to_stan.R
\name{format_covariates}
\alias{format_covariates}
\title{Format covariates data for stan estimation}
\usage{
format_covariates(t, H, n, covariates)
}
\arguments{
\item{t}{T: number of layers in the multiplex network}

\item{H}{number of pairs of layers}

\item{n}{number of actors}

\item{covariates}{list of covariates in the intermediate form:
covariate <- list(is_within = {T, F}, t_or_h, var_name, value)}
}
\value{
A list of covariates formatted to be estimated via multiplex_p2.stan
}
\description{
Format covariates data for stan estimation
}
