% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdata_to_stan.R
\name{format_network}
\alias{format_network}
\title{Format network data for stan estimation}
\usage{
format_network(M)
}
\arguments{
\item{M}{list of the observed networks in adjacency matrix form}
}
\value{
list of formatted network related data to feed to Stan of the form list(n=n, N=N, T=t, H=H, layer_pairs=pairs,y_obs=y, K=K, N_obs = length(y),obs_idx = 1:N)
}
\description{
Format network data for stan estimation
}
