% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_dyads}
\alias{get_dyads}
\title{Get pairs given total number of actors/layers}
\usage{
get_dyads(n)
}
\arguments{
\item{n}{numeric: number of actors (or number of layers)}
}
\value{
list of vectors of size 2: indices of the pairs of actors (layers) in the networks (dyads)
}
\description{
Get pairs given total number of actors/layers
}
