% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{gof_plot}
\alias{gof_plot}
\title{Plot Goodness of Fit}
\usage{
gof_plot(
  sim_stat_res,
  net_lab,
  center = FALSE,
  scale = FALSE,
  violin = TRUE,
  key = NULL,
  perc = 0.05,
  position = 4,
  fontsize = 12,
  ...
)
}
\arguments{
\item{sim_stat_res}{A list of simulated network statistics. Each element of the list should be a list with two elements: \code{stat_sim} and \code{stat_obs}. \code{stat_sim} is a matrix of simulated network statistics, and \code{stat_obs} is a matrix of observed network statistics.}

\item{net_lab}{The label of the network.}

\item{center}{Logical indicating whether to center the statistics.}

\item{scale}{Logical indicating whether to scale the statistics.}

\item{violin}{Logical indicating whether to plot violin plots.}

\item{key}{Optional vector of labels for the x-axis.}

\item{perc}{The percentage of simulations to use for the confidence intervals.}

\item{position}{The position of the observation labels.}

\item{fontsize}{The font size for the plot.}

\item{...}{Additional arguments to be passed to the plotting functions.}
}
\value{
None.
}
\description{
This function plots the goodness of fit for a Mp2 model
}
