% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ifelse_helper}
\alias{ifelse_helper}
\title{ifelse_helper Function}
\usage{
ifelse_helper(x, yes, no)
}
\arguments{
\item{x}{A logical condition.}

\item{yes}{The value to return if the condition is TRUE.}

\item{no}{The value to return if the condition is FALSE.}
}
\value{
The value of \code{yes} if the condition is TRUE, and the value of \code{no} otherwise.
}
\description{
This function is a helper function that mimics the behavior of the \code{ifelse} function in R.
It takes a logical condition \code{x} and returns \code{yes} if the condition is TRUE, and \code{no} otherwise.
}
