% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{Indegree_distribution}
\alias{Indegree_distribution}
\title{Calculate the indegree distribution of a single network}
\usage{
Indegree_distribution(single_net, levels = 0:8, cumulative = TRUE)
}
\arguments{
\item{single_net}{The single network for which to calculate the indegree distribution.}

\item{levels}{The levels of indegree to consider. Default is 0:8.}

\item{cumulative}{Logical indicating whether to calculate the cumulative distribution. Default is TRUE.}
}
\value{
The indegree distribution of the single network.
}
\description{
This function calculates the indegree distribution of a single network.
The indegree distribution is a measure of the number of outgoing edges from each node in the network.
}
