% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_valid}
\alias{is_valid}
\title{Check if the input is valid for a given type.}
\usage{
is_valid(x, type, n)
}
\arguments{
\item{x}{The input to be checked.}

\item{type}{The type of input to be checked.}

\item{n}{The number of actors in the network.}
}
\value{
TRUE if the input is valid, FALSE otherwise.
}
\description{
This function checks if the input is valid for a given type, such as "dep_net", "dyad_covar", or "actor_covar".
For "dep_net" type, the input should be a list of matrices with binary values.
For "dyad_covar" type, the input should be a list of numeric matrix with no missing values.
For "actor_covar" type, the input should be a numeric data frame with no missing values.
}
