% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{LKJ_decomp}
\alias{LKJ_decomp}
\title{Generate a correlation matrix using the LKJ decomposition method}
\usage{
LKJ_decomp(eta, t)
}
\arguments{
\item{eta}{Parameter for the LKJ distribution}

\item{t}{number if networks}
}
\value{
A correlation matrix
}
\description{
Generate a correlation matrix using the LKJ decomposition method
}
