% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{matrix_to_dyads_1d}
\alias{matrix_to_dyads_1d}
\title{Obtain dyad outcomes from a single network layer}
\usage{
matrix_to_dyads_1d(M)
}
\arguments{
\item{M}{a single adj matrices representing one layer of the multiplex network}
}
\value{
return a vector y of size N the dyad-wise outcome y in {1,2,3,4}^N
}
\description{
Obtain dyad outcomes from a single network layer
}
