% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{matrix_to_dyads_nd}
\alias{matrix_to_dyads_nd}
\title{Obtain dyad outcomes from network matrices}
\usage{
matrix_to_dyads_nd(Ms)
}
\arguments{
\item{Ms}{list of adj matrices representing the multiplex network}
}
\value{
return a vector y of size N, such that y in {1,...,2^(2*t)}^N representing the outcomes on a pair of dyads.
}
\description{
Obtain dyad outcomes from network matrices
}
