% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Mp2Model}
\alias{Mp2Model}
\title{Mp2Model function}
\usage{
Mp2Model(dep_net, dyad_covar = NULL, actor_covar = NULL, ...)
}
\arguments{
\item{dep_net}{A named list of matrices representing the dependent multiplex network (outcome), values are binary or NA.}

\item{dyad_covar}{A list of matrices representing the dyadic covariates (optional), logical or numeric, but not NA.}

\item{actor_covar}{A data frame representing the actor covariates (optional), logical or numeric, but not NA.}
}
\value{
A new Mp2Model object representing the empty model.
}
\description{
This function creates an empty model of the class Mp2Model.
}
\details{
This function creates a default empty model for multip2 analysis. It takes in the dependent network, dyadic covariates, and actor covariates (optional) as input. The function checks the validity of the input arguments and returns a new Mp2Model object with the specified attributes.
}
\examples{
#not run
#model <- Mp2Model(dep_net, dyad_covar, actor_covar)
}
