% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multip2-package.R
\docType{package}
\name{multip2-package}
\alias{multip2}
\alias{multip2-package}
\title{multip2: Bayesian Multiplex Mixed-effects Network Model}
\description{
Multiplex extension to the p2 (van Duijn et al., 2004) model which includes cross-layer dyadic dependencies as fixed effects and actor-specific dependencies as random effects. The influence of covariates are also considered in the analysis of cross-sectional, directed binary multiplex network data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/annihong/multip2}
  \item Report bugs at \url{https://github.com/annihong/multip2/issues}
}

}
\author{
\strong{Maintainer}: Anni Hong \email{annihong@andrew.cmu.edu} (\href{https://orcid.org/0000-0001-6562-9334}{ORCID})

}
\keyword{internal}
