% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MultiP2Fit}
\alias{MultiP2Fit}
\title{Format covariates data for stan estimation}
\usage{
MultiP2Fit(
  network_data = list(),
  actor_data = data.frame(),
  outcome = "",
  sender_covar = "",
  receiver_covar = "",
  density_covar = "",
  recip_covar = "",
  cross_density_covar = "",
  cross_recip_covar = ""
)
}
\arguments{
\item{network_data}{list of p + t matrices of dim n x n : network outcomes and network covariates, p is the number of network covariates and t is the number of layers in the multiplex network outcome.}

\item{actor_data}{n x p data frame: p covariates of length of the number of actors (n)}

\item{outcome}{vector: names of the layers of networks in the network_data as the outcome multiplex network}
}
\value{
a MultiP2Fit object
}
\description{
Format covariates data for stan estimation
}
\examples{
networks = replicate(4, matrix(data=1:9, nrow=3), simplify = FALSE)
names(networks) <-  c("network1", "network2", "network_covariate1",      "network_covariate2")
MultiP2Fit(outcome = c("network1", "network2"), 
network_data = networks, 
actor_data = data.frame(actor_attr1=1:3, actor_attr2=1:3))
}
