% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{multiplex_gof_baseline}
\alias{multiplex_gof_baseline}
\title{multiplex_gof_baseline}
\usage{
multiplex_gof_baseline(
  dep_net,
  sim_nets,
  network_statistics,
  return_data,
  descriptive_labels = NULL...
)
}
\arguments{
\item{dep_net}{The observed multiplex network.}

\item{sim_nets}{A list of simulated multiplex networks.}

\item{network_statistics}{A list of network statistics to compute for each network.}

\item{return_data}{A logical value indicating whether to return the computed statistics.}

\item{descriptive_labels}{Optional labels for the x-axis of the plot.}
}
\value{
If return_data is TRUE, a list containing the computed statistics for the simulated and observed networks.
If return_data is FALSE, a ggplot object representing the comparison plot.
}
\description{
This function compares the observed and simulated basic multiplex statistics using boxplots and scatter plots.
}
