% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.Mp2Model}
\alias{print.Mp2Model}
\title{Print Mp2Model Object}
\usage{
\method{print}{Mp2Model}(model_obj, ...)
}
\arguments{
\item{model_obj}{An Mp2Model object.}

\item{...}{Additional arguments to be passed to the print function.}
}
\value{
None
}
\description{
This function prints information about an Mp2Model object.
}
