% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{prior_simulated_network_checks}
\alias{prior_simulated_network_checks}
\title{prior_simulated_network_checks}
\usage{
prior_simulated_network_checks(
  simulated_networks,
  descriptive_stats_func = descriptive_stats_prior,
  stats_lab = c("density", "reciprocity", "transitivity")
)
}
\arguments{
\item{simulated_networks}{A list of simulated networks.}

\item{descriptive_stats_func}{The function used to calculate descriptive statistics for each network, default is descriptive_stats_prior (calculates density, reciprocity, and transitivity).}

\item{stats_lab}{A character vector specifying the labels for the descriptive statistics.}
}
\description{
This function performs predictive checks on simulated networks.
}
\details{
This function iterates over each layer of simulated networks and calculates descriptive statistics for each uniplex network.
It then plots histograms of the descriptive statistics for each layer and displays them using grid.arrange.
The function also allows customization of the descriptive statistics function and the labels for the statistics.
}
\examples{
#prior_simulated_network_checks(simulated_networks, descriptive_stats_func = descriptive_stats_prior, stats_lab = c("density", "reciprocity", "transitivity"))

}
