% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sample_prior}
\alias{sample_prior}
\title{sample_prior function}
\usage{
sample_prior(n, t, mu_0, rho_0, cross_mu_0, cross_rho_0, Sigma)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{t}{The number of networks.}

\item{mu_0}{The prior mean for mu.}

\item{rho_0}{The prior mean for rho.}

\item{cross_mu_0}{The prior mean for cross_mu.}

\item{cross_rho_0}{The prior mean for cross_rho.}

\item{Sigma}{The covariance matrix.}
}
\value{
A list containing the generated samples for mu, rho, cross_mu, cross_rho, C, and Sigma.
}
\description{
This function generates samples from the priors.
}
