% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sigma_cauchy}
\alias{sigma_cauchy}
\title{Generate a vector of random numbers from a Cauchy distribution}
\usage{
sigma_cauchy(k, t)
}
\arguments{
\item{k}{Scale parameter of the Cauchy distribution}

\item{t}{Number of networks}
}
\value{
A vector of random numbers
}
\description{
Generate a vector of random numbers from a Cauchy distribution
}
