% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sigma_folded_normal}
\alias{sigma_folded_normal}
\title{Generate a vector of random numbers from a folded normal distribution}
\usage{
sigma_folded_normal(mean, sd, t)
}
\arguments{
\item{mean}{Mean of the folded normal distribution}

\item{sd}{Standard deviation of the folded normal distribution}

\item{t}{Number of networks}
}
\value{
A vector of random numbers
}
\description{
Generate a vector of random numbers from a folded normal distribution
}
