% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sigma_inver_gamma}
\alias{sigma_inver_gamma}
\title{Generate a vector of random numbers from an inverse gamma distribution}
\usage{
sigma_inver_gamma(shape, scale, t)
}
\arguments{
\item{shape}{Shape parameter of the inverse gamma distribution}

\item{scale}{Scale parameter of the inverse gamma distribution}

\item{t}{Number of networks}
}
\value{
A vector of random numbers
}
\description{
Generate a vector of random numbers from an inverse gamma distribution
}
