% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{Sigma_LJK}
\alias{Sigma_LJK}
\title{Generate a covariance matrix using the LKJ decomposition method}
\usage{
Sigma_LJK(sigma, L_corr)
}
\arguments{
\item{sigma}{A vector of standard deviations}

\item{L_corr}{A correlation matrix}
}
\value{
A covariance matrix
}
\description{
Generate a covariance matrix using the LKJ decomposition method
}
