% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_random_effects}
\alias{sim_random_effects}
\title{Given number of actors, total number of networks, and the var-cov matrix
return a list of sender (alpha) \in R^{n x T} and receiver effects (beta)}
\usage{
sim_random_effects(n_actor, T_network, Sigma, supplied_C = FALSE)
}
\arguments{
\item{n_actor}{Number of actors}

\item{T_network}{Total number of networks}

\item{Sigma}{Variance-covariance matrix}

\item{supplied_C}{Boolean indicating whether the var-cov matrix is supplied or generated}
}
\value{
A list containing the sender effects (alpha) and receiver effects (beta)
}
\description{
Given number of actors, total number of networks, and the var-cov matrix
return a list of sender (alpha) \in R^{n x T} and receiver effects (beta)
}
