% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{simulated_network_checks}
\alias{simulated_network_checks}
\title{simulated_network_checks Function}
\usage{
simulated_network_checks(
  sim_nets,
  model_obj,
  network_statistics,
  layer_lab = NULL,
  cumulative = TRUE,
  return_data = FALSE,
  center = FALSE,
  scale = FALSE,
  violin = TRUE,
  key = NULL,
  perc = 0.05,
  position = 4,
  fontsize = 12,
  ...
)
}
\arguments{
\item{model_obj}{The fitted model object.}

\item{network_statistics}{The network statistics function to be used for the checks.}

\item{layer_lab}{The labels of the network layers to be used. If NULL, all network layers are used.}

\item{cumulative}{Logical indicating whether to calculate cumulative statistics.}

\item{return_data}{Logical indicating whether to return the simulated network statistics.}

\item{center}{Logical indicating whether to center the statistics.}

\item{scale}{Logical indicating whether to scale the statistics.}

\item{violin}{Logical indicating whether to plot the statistics using violin plots.}

\item{key}{The key for the violin plots.}

\item{perc}{The percentile for the key.}

\item{position}{The position of the key.}

\item{fontsize}{The font size of the plot.}

\item{...}{Additional arguments to be passed to the network statistics function.}

\item{num_sim}{The number of simulated networks to generate.}
}
\value{
If return_data is TRUE, a list of simulated network statistics is returned.
Otherwise, the function plots the simulated network statistics.
}
\description{
This function performs predictive checks on a fitted model using simulated networks.
}
\details{
This function checks the goodness-of-fit of a fitted model by comparing the observed network statistics
with the statistics calculated from the simulated networks. It supports both multiplex statistics and
single layer statistics.
}
