% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdata_to_stan.R
\name{stan_data}
\alias{stan_data}
\title{The main function to convert user network input to Stan data}
\usage{
stan_data(M, covariates)
}
\arguments{
\item{M}{list of the observed networks in adjacency matrix form}

\item{covariates}{intermediate form of the covariates: covariate <- list(is_within = {T, F}, t_or_h, var_name, value)}
}
\value{
a list in the format for Stan code
}
\description{
The main function to convert user network input to Stan data
}
