% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.Mp2Model}
\alias{summary.Mp2Model}
\title{This function prints the summary of a Mp2Model object.}
\usage{
summary.Mp2Model(model_obj, ...)
}
\arguments{
\item{model_obj}{A Mp2Model object.}

\item{...}{Additional arguments to be passed to the print function.}
}
\value{
None
}
\description{
This function prints the summary of a Mp2Model object.
}
