% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{triad_census}
\alias{triad_census}
\title{Triad Census}
\usage{
triad_census(single_net, levels = 1:16)
}
\arguments{
\item{single_net}{The adjacency matrix of the network.}

\item{levels}{The levels of triad types to include in the census.
Default is 1:16, which includes all triad types.}
}
\value{
A vector containing the counts of each triad type specified by the levels parameter.
}
\description{
This function calculates the triad census for a given network.
}
