% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{update_covar}
\alias{update_covar}
\title{Add covariates to the model}
\usage{
update_covar(
  model_obj,
  layer_1,
  layer_2 = NULL,
  density = NULL,
  reciprocity = NULL,
  cross_density = NULL,
  cross_reciprocity = NULL,
  sender = NULL,
  receiver = NULL
)
}
\arguments{
\item{model_obj}{The model object to which the covariate will be added.}

\item{covar_lab}{The label of the covariate.}

\item{param}{The parameter associated with the covariate.}

\item{layer_lab}{The label of the layer (for within-layer covariates) or pair (for cross-layer covariates) to which the covariate belongs. If not specified, the covariate will be added to all layers or pairs.}

\item{mean_prior}{The mean prior for the covariate. If not specified, no prior will be used.}

\item{sd_prior}{The standard deviation prior for the covariate. If not specified, no prior will be used.}
}
\value{
The updated model object with the added covariate.
}
\description{
This function adds covariates to the model object. It supports different types of covariates, such as within-layer covariates and cross-layer covariates. The function checks the type of covariate and adds it to the appropriate section of the model object.
}
