% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{y_nd_ij_to_y_1d_ij}
\alias{y_nd_ij_to_y_1d_ij}
\title{Convert Dyadic Data for a Single Dyad to Multiple Networks}
\usage{
y_nd_ij_to_y_1d_ij(y_nd_ij, total)
}
\arguments{
\item{y_nd_ij}{Dyadic data for a single dyad in {1,...,2^(2*total)}.}

\item{total}{The total number of networks.}
}
\value{
A list of \code{total} dyadic data for the dyad in {1,2,3,4}.
}
\description{
Given dyadic data \code{y_nd_ij} for a single dyad and the total number of networks \code{total},
this function returns a list of \code{total} dyadic data for the dyad in {1,2,3,4}.
}
