% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract draws from the (prior) posterior of a fitted stan object}
\usage{
extract_draws(fit, parameter)
}
\arguments{
\item{fit}{rstan fit object: fitted stan object}

\item{parameter}{string: name of the desired parameters, y_tilde is the simulated outcome of the model}
}
\value{
the (prior) posterior draws of the specified parameters
}
\description{
Extract draws from the (prior) posterior of a fitted stan object
}
