% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{extract_network_draws}
\alias{extract_network_draws}
\title{get draws function}
\usage{
extract_network_draws(model_obj, sim_num, network_type = "adj")

extract_network_draws(model_obj, sim_num, network_type = "adj")
}
\arguments{
\item{model_obj}{An object of class "Mp2Model"}

\item{sim_num}{integer: number of simulations to extract, counting from the tail of the posterior draws}

\item{network_type}{string: type of network to return. Options are "adj" for adjacency matrix (default), "igraph" for igraph object, or "network" for network object.}
}
\value{
draws in the speficied format

A list of length \code{sim_num}. Each element of the list is another list of \code{t} networks. These networks represent the (prior) posterior draws of the simulated network outcome, in the specified format.
}
\description{
This function obtains the draws from the fitted Stan model of the Mp2Model class.
}
