% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fit.Mp2Model}
\alias{fit.Mp2Model}
\title{Fit Mp2Model}
\usage{
\method{fit}{Mp2Model}(
  model_obj,
  chains = 4,
  iter = 200,
  warmup = floor(iter/2),
  thin = 1,
  seed = sample.int(.Machine$integer.max, 1),
  prior_sim = FALSE,
  network_sim = TRUE,
  mc.cores = parallel::detectCores(),
  auto_write = TRUE,
  stan_file = "multiplex_p2_low_mem.stan",
  fit = TRUE,
  ...
)
}
\arguments{
\item{model_obj}{The Mp2Model object.}

\item{chains}{The number of chains to run in parallel (default: 4), see rstan::stan() documentation for more details.}

\item{iter}{The total number of iterations (default: 200), see rstan::stan() documentation for more details.}

\item{warmup}{The number of warmup iterations (default: floor(iter/2)), see rstan::stan() documentation for more details.}

\item{thin}{The thinning interval (default: 1), see rstan::stan() documentation for more details.}

\item{seed}{The random seed (default: a random integer), see rstan::stan() documentation for more details.}

\item{prior_sim}{Whether to simulate from the prior distribution (default: FALSE).}

\item{mc.cores}{The number of CPU cores to use for parallel computing (default: all available cores by calling parallel::detectCores()), see rstan::stan() documentation for more details.}

\item{auto_write}{Whether to automatically write compiled Stan models to disk (default: TRUE), see rstan::stan() documentation for more details.}

\item{stan_file}{The name of the Stan file (default: "multiplex_p2_low_mem.stan").}

\item{...}{Additional arguments passed to rstan::stan(), see rstan::stan() documentation for more details.}
}
\value{
The updated Mp2Model object with fit results.
}
\description{
This function fits the Mp2Model using Stan.
}
\examples{
#model <- Mp2Model(...)
#model <- fit(model)
#summary(model)
}
