% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fit.Mp2Model}
\alias{fit.Mp2Model}
\title{fit function for "Mp2Model" class}
\usage{
\method{fit}{Mp2Model}(
  model_obj,
  chains = 4,
  iter = 200,
  warmup = floor(iter/2),
  thin = 1,
  seed = sample.int(.Machine$integer.max, 1),
  stan_file = "multiplex_p2_low_mem.stan",
  prior_sim = FALSE,
  mc.cores = parallel::detectCores(),
  auto_write = TRUE,
  ...
)
}
\arguments{
\item{model_obj}{An object of class "Mp2Model"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
fit function for "Mp2Model" class
}
