% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fit}
\alias{fit}
\title{fit function}
\usage{
fit(model_obj, ...)
}
\arguments{
\item{model_obj}{An object of class "Mp2Model"}
}
\value{
An object of class "Mp2Model" with the fitted Stan model and parameter labels
}
\description{
This function performs the estimation of the multiplex P2 model using Stan.
}
