% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdata_to_stan.R
\name{format_covariates}
\alias{format_covariates}
\title{Format covariates data for stan estimation}
\usage{
format_covariates(t, H, n, covariates)
}
\arguments{
\item{t}{T: number of layers in the multiplex network}

\item{H}{number of pairs of layers}

\item{n}{number of actors}

\item{covars}{}
}
\value{
A list of covariates formatted to be estimated via multiplex_p2.stan
}
\description{
Format covariates data for stan estimation
}
\examples{
format_covariates(2, 1, 30, covariates=list())
}
