% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MultiP2Fit}
\alias{MultiP2Fit}
\title{Format covariates data for stan estimation}
\usage{
MultiP2Fit(
  network_data,
  outcome,
  actor_data = data.frame(),
  senders_covar = NULL,
  receivers_covar = NULL,
  density_covar = NULL,
  reciprocity_covar = NULL,
  cross_density_covar = NULL,
  cross_reciprocity_covar = NULL,
  custom_covars = NULL,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2),
  thin = 1,
  seed = sample.int(.Machine$integer.max, 1),
  stan_file = "multiplex_p2.stan",
  prior_sim = FALSE
)
}
\arguments{
\item{network_data}{list of p + t matrices of dim n x n : network outcomes and network covariates, p is the number of network covariates and t is the number of layers in the multiplex network outcome.}

\item{outcome}{A character vector of the names of the layers of networks in the network_data as the outcome multiplex network}

\item{actor_data}{n x p data frame: p covariates of length of the number of actors (n)}

\item{senders_covar}{A character vector indicating the names of the covariates to include for the senders.}

\item{receivers_covar}{A character vector indicating the names of the covariates to include for the receivers.}

\item{density_covar}{A character vector indicating the names of the covariates to include for the density.}

\item{reciprocity_covar}{A character vector indicating the names of the covariates to include for the reciprocity.}

\item{cross_density_covar}{A character vector indicating the names of the covariates to include for the cross-layer density.}

\item{cross_reciprocity_covar}{A character vector indicating the names of the covariates to include for the cross-layer reciprocity.}

\item{custom_covars}{A list of custom covariates to include in the model.}

\item{chains}{An integer indicating the number of Markov chains to run.}

\item{iter}{An integer indicating the number of iterations per chain.}

\item{warmup}{An integer indicating the number of warmup iterations per chain.}

\item{thin}{An integer indicating the thinning rate for the chains.}

\item{seed}{An integer indicating the random seed for the chains.}

\item{stan_file}{A character string indicating the name of the stan file to use, default is "multiplex_p2.stan".}
}
\value{
a MultiP2Fit object

A list containing the fitted Stan model and parameter labels.
}
\description{
Format covariates data for stan estimation
}
