% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_checks.R
\name{prior_simulated_network_checks}
\alias{prior_simulated_network_checks}
\title{Perform checks on simulated networks}
\usage{
prior_simulated_network_checks(
  simulated_networks,
  descriptive_stats_func = descriptive_stats_prior,
  stats_lab = c("density", "reciprocity", "transitivity")
)
}
\arguments{
\item{simulated_networks}{A list of simulated networks. Each element of the list should be an list of adjacency matrix (a multiplex network).}

\item{stats_lab}{A character vector of labels for the statistics calculated by \code{descriptive_stats}. The default is \code{c("density", "reciprocity", "transitivity")}.}

\item{descriptive_stats}{A function that calculates descriptive statistics for a network. This function should take an igraph object as input and return a numeric vector of statistics.}
}
\value{
A ggplot2 object. This plot shows histograms of the calculated statistics for the simulated networks.
}
\description{
This function takes a list of simulated networks, calculates descriptive statistics for each network,
and returns a ggplot2 object of histograms of these statistics.
}
\examples{
# TODO: Add examples of usage
}
