% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{update_covar}
\alias{update_covar}
\title{Update Covariate}
\usage{
update_covar(
  model_obj,
  layer_1,
  layer_2 = NULL,
  density = NULL,
  reciprocity = NULL,
  cross_density = NULL,
  cross_reciprocity = NULL,
  sender = NULL,
  receiver = NULL
)
}
\arguments{
\item{model_obj}{An object of the class Mp2Model to update the covariates for.}

\item{layer_1}{The first layer of the network to specify the covariates for.}

\item{layer_2}{The second layer of the network to specify the covariates for (optional, used for specifying cross-layer covariates).}

\item{density}{A vector of covariate labels for density in layer_1.}

\item{reciprocity}{A vector of covariate labels for reciprocity in layer_1.}

\item{cross_density}{A vector of covariate labels for cross-layer density in layer_1 x layer_2 (only when layer_2 is not null).}

\item{cross_reciprocity}{A vector of covariate labels for cross-layer reciprocity in layer_1 x layer_2 (only when layer_2 is not null).}

\item{sender}{A vector of covariate labels for sender effect in layer_1.}

\item{receiver}{A vector of covariate labels for receiver effect in layer_1.}
}
\value{
The updated Mp2Model with the new covariates.
}
\description{
This function updates the covariates in a given model object based on the specified parameters.
}
